threading bug fix
